Add bf16 support for model training and evaluation in constants and train scripts

+++ wisper/train.py
@@ -86,7 +86,10 @@ def train(resume_from_checkpoint: Optional[str] = None):
         return {"wer": wer}
 
     t0 = time.perf_counter()
-    model = WhisperForConditionalGeneration.from_pretrained(model_source)
+    model = WhisperForConditionalGeneration.from_pretrained(
+        model_source,
+        torch_dtype=torch.bfloat16,
+    )
     model.generation_config.language = LANGUAGE
     model.generation_config.task = TASK
     model.generation_config.forced_decoder_ids = None
